requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
facenet-pytorch==2.6.0
pandas==2.0.3
openpyxl==3.1.2
Pillow>=10.2.0,<10.3.0
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.3.0
torch>=2.2.0,<2.3.0
torchvision>=0.17.0,<0.18.0
albumentations==1.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
numpy==1.26.4